Deprecate CoroutineWorker, enable the new Memory Model
