+++ gradle.properties
@@ -1,5 +1,6 @@
-VERSION=0.8.2
+VERSION=0.9.0
 
 kotlin.mpp.enableGranularSourceSetsMetadata=true
 kotlin.mpp.enableCompatibilityMetadataVariant=true
 kotlin.mpp.stability.nowarn=true
+kotlin.native.binary.memoryModel=experimental
